Refactor bot_utils out of health-checker
